move database/table name variables to infrastructure.py
