+++ source/databricks/calculation-engine/package/constants/__init__.py
@@ -17,6 +17,3 @@ from .basis_data_colname import BasisDataColname
 from .partition_key_name import PartitionKeyName
 from .result_table_col_name import ResultTableColName
 
-OUTPUT_DATABASE_NAME = "wholesale_output"
-ENERGY_RESULT_TABLE_NAME = "result"
-TEST = ""
